drm: Enable bridges asynchronously as early as possible
Currently, the display takes a while to power on because there is a long
delay between when the first commit is received to power on the display
and when the display is actually powered on. Additionally, the display
power on sequence blocks the display driver from proceeding with commit
dispatch, which adds even more latency to the display power on sequence.

In order to make the display turn on faster, enable the drm bridges
asynchronously as soon as the first commit is received when the display is
powered off. To ensure that this doesn't introduce races within the display
drivers utilizing drm, the regular drm_bridge_pre_enable and
drm_bridge_enable functions are turned into stubs that block until the
kthread worker responsible for enabling the bridges finishes doing so.

Change-Id: Ifc6b93cac86445c9a3c89260a5c06485af0c7978
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: YousefAlgadri <yusufgadrie@gmail.com>