clk: qcom: clk-cpu-osm: Don't sleep inside the global clk lock
The clk framework uses a global mutex for all clock operations. This
means that when one clock user has the lock held, no other clocks can be
changed during that time. This can result in slow clock-rate propagation
to unrelated clocks.

To remedy this, only sleep within a per-cluster mutex lock. This not
only allows different clusters to quickly update their clock rates
independent of one another, but also prevents stalls when non-CPU clocks
are updated, since the global clk lock will not be held for a long time.

Change-Id: I3055a1358421c73df5df9892ccf5ebd6d852fe29
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Yousef Algadri <yusufgadrie@gmail.com>